WIP: Accurate seek for MP3
MP3 doesn't provide timestamps or high-resolution indices for
seeking, so to achieve high seek accuracy in large files we
need to decode frames and accumulate the elapsed time until we
hit the target time.

This is only a demo implementation and could use improvement.
Eg. by building a seek index when decoding, it will be faster
to seek to a part of the file that was previously decoded.

Change-Id: I00518460c309f4ce36f930085e57ec47c4165a0e