pwrmgmt: sleep timer: don't shut down while charging
This mirrors the behavior of idle poweroff,
which inhibits shutdown as long as a charger
is plugged in, even if a device is capable
of powering off while charging.

Since usb_inserted() already checks for USB_POWERED,
certain devices with the ability to power off while
charging, already exhibit this behavior when using
the sleep timer anyway.

Change-Id: I35ed4b542a8a4df06a34395c85f4d37fc1d2ce53